Less ugly.
