+++ middlewares/jwt/middleware.go
@@ -1,0 +1,70 @@
+package jwt
+
+import (
+	"encoding/json"
+	"fmt"
+	"net/http"
+
+	owner "github.com/factorysh/microdensity/claims"
+	"github.com/factorysh/microdensity/httpcontext"
+	"github.com/getsentry/sentry-go"
+)
+
+// Auth will ensure JWT token is valid
+func (j *JWTAuthenticator) Middleware() func(next http.Handler) http.Handler {
+	return func(next http.Handler) http.Handler {
+		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+
+			// if user was auth using oauth
+			isOAuth, err := httpcontext.GetIsOAuth(r)
+			if err == nil && isOAuth {
+				next.ServeHTTP(w, r)
+				return
+			}
+
+			token, err := httpcontext.GetJWT(r)
+			if err != nil {
+				w.WriteHeader(http.StatusUnauthorized)
+				return
+			}
+
+			j.handleJWTAuth(token, w, r, next)
+		})
+	}
+}
+
+func (j *JWTAuthenticator) handleJWTAuth(token string, w http.ResponseWriter, r *http.Request, next http.Handler) {
+	t, err := j.ParseAndValidate(token)
+
+	if err != nil {
+		fmt.Println(err)
+		w.WriteHeader(http.StatusInternalServerError)
+		return
+	}
+
+	// owner claims
+	var claims owner.Claims
+	err = json.Unmarshal(t.RawClaims(), &claims)
+	if err != nil {
+		fmt.Println(err)
+		w.WriteHeader(http.StatusUnauthorized)
+		return
+	}
+
+	err = claims.Validate()
+	if err != nil {
+		fmt.Println(err)
+		w.WriteHeader(http.StatusBadRequest)
+		return
+	}
+
+	if hub := sentry.GetHubFromContext(r.Context()); hub != nil {
+		hub.WithScope(func(scope *sentry.Scope) {
+			scope.SetExtra("jwt", claims)
+		})
+	}
+
+	ctx := claims.ToCtx(r.Context())
+
+	next.ServeHTTP(w, r.WithContext(ctx))
+}
