+++ mockup/gitlab.go
@@ -2,6 +2,8 @@ package mockup
 
 import (
 	"crypto/rsa"
+	"crypto/sha1"
+	"encoding/base64"
 	"encoding/json"
 	"net/http"
 
@@ -9,13 +11,19 @@ import (
 )
 
 func GitlabJWK(public *rsa.PublicKey) http.Handler {
+	h := sha1.New()
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		kid := base64.RawURLEncoding.EncodeToString(h.Sum(public.N.Bytes()))[:16]
 		j := jose.JSONWebKey{
-			Key:   public,
-			Use:   "sig",
-			KeyID: "someID",
+			Key:       public,
+			Use:       "sig",
+			KeyID:     kid,
+			Algorithm: "RS256",
 		}
-		err := json.NewEncoder(w).Encode(j)
+		jwks := jose.JSONWebKeySet{
+			Keys: []jose.JSONWebKey{j},
+		}
+		err := json.NewEncoder(w).Encode(jwks)
 		if err != nil {
 			panic(err)
 		}
